requirements.txt
logging
numpy
Pillow
scikit-image